+++ tools/integration_tests/local_file/local_file_test.go
@@ -29,6 +29,7 @@ import (
 	"github.com/googlecloudplatform/gcsfuse/tools/integration_tests/util/mounting/dynamic_mounting"
 	"github.com/googlecloudplatform/gcsfuse/tools/integration_tests/util/mounting/only_dir_mounting"
 	"github.com/googlecloudplatform/gcsfuse/tools/integration_tests/util/mounting/static_mounting"
+	"github.com/googlecloudplatform/gcsfuse/tools/integration_tests/util/operations"
 	"github.com/googlecloudplatform/gcsfuse/tools/integration_tests/util/setup"
 )
 
@@ -41,6 +42,33 @@ var (
 	storageClient *storage.Client
 	ctx           context.Context
 )
+
+////////////////////////////////////////////////////////////////////////
+// Helpers
+////////////////////////////////////////////////////////////////////////
+
+func WritingToLocalFileShouldNotWriteToGCS(ctx context.Context, storageClient *storage.Client,
+	fh *os.File, testDirName, fileName string, t *testing.T) {
+	operations.WriteWithoutClose(fh, client.FileContents, t)
+	client.ValidateObjectNotFoundErrOnGCS(ctx, storageClient, testDirName, fileName, t)
+}
+
+func NewFileShouldGetSyncedToGCSAtClose(ctx context.Context, storageClient *storage.Client,
+	testDirPath, fileName string, t *testing.T) {
+	// Create a local file.
+	_, fh := client.CreateLocalFileInTestDir(ctx, storageClient, testDirPath, fileName, t)
+
+	// Writing contents to local file shouldn't create file on GCS.
+	testDirName := client.GetDirName(testDirPath)
+	WritingToLocalFileShouldNotWriteToGCS(ctx, storageClient, fh, testDirName, fileName, t)
+
+	// Close the file and validate if the file is created on GCS.
+	client.CloseFileAndValidateContentFromGCS(ctx, storageClient, fh, testDirName, fileName, client.FileContents, t)
+}
+
+////////////////////////////////////////////////////////////////////////
+// TestMain
+////////////////////////////////////////////////////////////////////////
 
 func TestMain(m *testing.M) {
 	setup.ParseSetUpFlags()
